Add files via upload
Guess the Number Game" project is a simple Java console application that generates a random number between 1 and 10 and allows the user to guess the number. It provides feedback to the user on whether the guess is too high or too low, and continues to prompt the user until they correctly guess the number. The program keeps track of the number of attempts it takes for the user to guess the number and outputs the result once the number is correctly guessed.